[#11] Separated error classes
